Target: Use float32.
This makes sure it doesn't get promoted to float64 or complex128.
We only use it with a 0.0 value anyway.

+++ jadapy/Target.py
@@ -1,8 +1,8 @@
-from numpy import float64
+from numpy import float32
 
-class Target(float64):
+class Target(float32):
     def __new__(cls, *args, **kwargs):
-        return float64.__new__(cls, *args, **kwargs)
+        return float32.__new__(cls, *args, **kwargs)
 
 
 SmallestMagnitude = Target(0.0)
